refactor: move the about component to a separate folder.
